Move session timeout check to middleware
Moved all session timeout processing to middleware, so it is accessible
by all views. This now checks for session data for both Plea and
Feedback journeys, as the feedback is now split into 2 screens.

Updated markup to include sessionTimeout JS variable used by
PromptOnChange module. The alternative message variable is also now a JS
var.

[MAPDEV446]
